+++ modules/iPhone/libs/GradientButtons/GradientButton.m
@@ -28,7 +28,7 @@
 {
     if (normalGradient == NULL)
     {
-        int locCount = [normalGradientLocations count];
+        NSUInteger locCount = [normalGradientLocations count];
         CGFloat locations[locCount];
         for (int i = 0; i < [normalGradientLocations count]; i++)
         {
